+++ histFactory/plots/plotTools.py
@@ -1,52 +1,14 @@
 import copy
+import os
+import sys
+import inspect
 
-from TTAnalysis import TT
-from ScaleFactors import *
+#### Get directory where scripts are stored to handle the import correctly
+scriptDir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
+sys.path.append(scriptDir + "/../../")
 
-#### Utility to join different cuts together (logical AND) ####
-
-def joinCuts(*cuts):
-    if len(cuts) == 0: 
-        return ""
-    elif len(cuts) == 1: 
-        return cuts[0]
-    else:
-        totalCut = "("
-        for cut in cuts:
-            cut = cut.strip().strip("&")
-            if cut == "":
-                continue
-            totalCut += "(" + cut + ")&&" 
-        totalCut = totalCut.strip("&") + ")"
-        return totalCut
-
-#### The IDs/... we want to run on ####
-
-# Default choice
-electronID = { TT.LepID.M: "M" }
-muonID = { TT.LepID.T: "T" }
-electronIso = { TT.LepIso.L: "L" }
-muonIso = { TT.LepIso.T: "T" }
-
-#electronID = { TT.LepID.L: "L", TT.LepID.M: "M" }
-#electronID = { TT.LepID.L: "L", TT.LepID.M: "M", TT.LepID.T: "T" }
-
-# Loose (for TFs)
-#electronID = { TT.LepID.L: "L" }
-#muonID = { TT.LepID.L: "L" }
-#electronIso = { TT.LepIso.L: "L" }
-#muonIso = { TT.LepIso.L: "L" }
-
-#myBWPs = { wp.first: wp.second for wp in TT.BWP.map }
-#myBWPs = { TT.BWP.L: "L", TT.BWP.M: "M" } 
-myBWPs = { TT.BWP.L: "L" } 
-
-myFlavours = [ "ElEl", "MuEl", "ElMu", "MuMu" ]
-#myFlavours = [ "ElEl" ]
-#myFlavours = [ "MuMu" ]
-#myFlavours = [ "ElMu", "MuEl" ]
-
-keepOnlySymmetricWP = False
+from common.TTAnalysis import *
+from common.ScaleFactors import *
 
 #### UTILITY TO GENERATE ALL THE NEEDED CATEGORIES ####
